GH-2: Reorganize activities into separate package.
